Added clang-tidy static analysis with CI integration
- Added .clang-tidy configuration with comprehensive checks
- Created tests/static_analysis/run_clang_tidy.sh for unified analysis
- Added GitHub Actions workflow for automated static analysis
- Fixed all clang-tidy warnings across kernel and userland code:
  * Resolved implicit widening conversions
  * Fixed parameter name consistency issues
  * Eliminated branch clone warnings
  * Optimized struct padding (vfs_file_t)
  * Corrected loop variable types
  * Added missing function declarations
  * Improved input validation in userland utilities
- Achieved 0 compilation errors and 0 real source warnings
- All recursive function warnings properly suppressed

+++ kernel/arch/riscv64/drivers/clint.c
@@ -9,8 +9,8 @@
 #define CLINT_REG_32(offset) ((volatile uint32_t *)(CLINT_BASE + (offset)))
 #define CLINT_REG_64(offset) ((volatile uint64_t *)(CLINT_BASE + (offset)))
 
-#define CLINT_MSIP_REG(hart) CLINT_REG_32(CLINT_MSIP_OFFSET + ((hart) * 4))
-#define CLINT_MTIMECMP_REG(hart) CLINT_REG_64(CLINT_MTIMECMP_OFFSET + ((hart) * 8))
+#define CLINT_MSIP_REG(hart) CLINT_REG_32(CLINT_MSIP_OFFSET + ((unsigned long)(hart) * 4UL))
+#define CLINT_MTIMECMP_REG(hart) CLINT_REG_64(CLINT_MTIMECMP_OFFSET + ((unsigned long)(hart) * 8UL))
 #define CLINT_MTIME_REG CLINT_REG_64(CLINT_MTIME_OFFSET)
 
 /* CSR access helpers */
